services/data-pipeline/requirements.txt
# Data Pipeline Dependencies
pyspark==3.1.0
kafka-python==2.0.2
psycopg2-binary==2.8.6
pandas==1.2.0
numpy==1.19.0
sqlalchemy==1.3.0
requests==2.25.0
pyyaml==5.3.1
boto3==1.17.0
redis==3.5.0
celery==4.4.0
